+++ UniGame.DOTweenRxExtensions/Runtime/unigame.dotweenrxextensions.runtime.asmdef
@@ -1,0 +1,17 @@
+{
+    "name": "unigame.dotweenextensions.runtime",
+    "references": [
+        "GUID:dd65fd2b6c82f454aad02c285f179b0c",
+        "GUID:560b04d1a97f54a4e82edc0cbbb69285",
+        "GUID:f51ebe6a0ceec4240a699833d6309b23"
+    ],
+    "includePlatforms": [],
+    "excludePlatforms": [],
+    "allowUnsafeCode": false,
+    "overrideReferences": false,
+    "precompiledReferences": [],
+    "autoReferenced": true,
+    "defineConstraints": [],
+    "versionDefines": [],
+    "noEngineReferences": false
+}
