+++ UniGame.DOTweenRxExtensions/Runtime/unigame.dotweenrxextensions.runtime.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: cfd40074e4fdde34e8eb3c8594715975
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
